refactor(PeoplePicker): separate PeoplePicker into separate Logic, Body & Buttons files
